+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## WIP
+
+### Improved
+- [Server] Faster code path for `pull` and `pull-many` on server.
+  [#322](https://github.com/juji-io/datalevin/issues/322)
+- [KV] Set default of `:max-readers` to 512.
+
 ## 0.9.22 (2025-03-18)
 
 ### Improved